refactor: split oversized frontend components into focused subcomponents
Addresses issue 3.1 from docs/code-quality-analysis.md.

ActionFeed.tsx (1,406 → 796 lines):
- Extract ActionCard component for individual action card rendering
- Extract ActionSearchDropdown with embedded ScoreTable for search/filter UI

VisualizationPanel.tsx (1,285 → 554 lines):
- Extract SldOverlay to dedicated component file (705 lines)
- Extract MemoizedSvgContainer to dedicated component file (48 lines)

CombinedActionsModal.tsx (777 → 397 lines):
- Extract ComputedPairsTable for the computed pairs tab
- Extract ExplorePairsTab for the explore/selection tab

useDiagrams.ts (767 → 693 lines):
- Extract useSldOverlay hook for SLD overlay state management

All 454 tests pass, zero TypeScript errors, zero lint warnings.
