+++ frontend/src/components/MemoizedSvgContainer.tsx
@@ -1,0 +1,48 @@
+// Copyright (c) 2025-2026, RTE (https://www.rte-france.com)
+// This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
+// If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+// SPDX-License-Identifier: MPL-2.0
+// This file is part of Co-Study4Grid a Power Grid Study tool Assistant Interface to help solve contigencies for a grid state under study.
+
+import React, { type RefObject } from 'react';
+import type { TabId } from '../types';
+
+// Prevents React from diffing massive SVG DOM trees on every parent render.
+// Uses replaceChildren(svgElement) instead of innerHTML to avoid the double-parse:
+//   OLD: string → XMLSerializer → string → browser parse (twice)
+//   NEW: DOMParser (in processSvg) → SVGSVGElement → replaceChildren (zero extra parse)
+interface SvgContainerProps {
+    svg: SVGSVGElement | string;
+    containerRef: RefObject<HTMLDivElement | null>;
+    display: string;
+    tabId: TabId;
+}
+
+const MemoizedSvgContainer = React.memo(({ svg, containerRef, display, tabId }: SvgContainerProps) => {
+    React.useLayoutEffect(() => {
+        const container = containerRef.current;
+        if (!container || !svg) return;
+
+        const start = performance.now();
+        if (svg instanceof SVGSVGElement) {
+            // DOM-reuse path: move the already-parsed element directly — no second parse
+            container.replaceChildren(svg);
+        } else {
+            // Fallback for plain string SVGs (e.g. SLD overlays not going through processSvg)
+            container.innerHTML = svg;
+        }
+        console.log(`[SVG] DOM injection for ${tabId} took ${(performance.now() - start).toFixed(2)}ms`);
+    }, [svg, containerRef, tabId]);
+
+    return (
+        <div
+            ref={containerRef}
+            className="svg-container"
+            id={`${tabId}-svg-container`}
+            style={{ display, width: '100%', height: '100%', overflow: 'hidden' }}
+        />
+    );
+});
+
+export default MemoizedSvgContainer;
