+++ frontend/src/components/SldOverlay.tsx
@@ -1,0 +1,705 @@
+// Copyright (c) 2025-2026, RTE (https://www.rte-france.com)
+// This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
+// If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+// SPDX-License-Identifier: MPL-2.0
+// This file is part of Co-Study4Grid a Power Grid Study tool Assistant Interface to help solve contigencies for a grid state under study.
+
+import React, { useState, useEffect, useRef } from 'react';
+import type { DiagramData, AnalysisResult, ActionDetail, VlOverlay, SldTab, SldFeederNode } from '../types';
+import { isCouplingAction } from '../utils/svgUtils';
+
+export interface SldOverlayProps {
+    vlOverlay: VlOverlay;
+    actionViewMode: 'network' | 'delta';
+    onOverlayClose: () => void;
+    onOverlaySldTabChange: (tab: SldTab) => void;
+    n1Diagram: DiagramData | null;
+    actionDiagram: DiagramData | null;
+    selectedBranch: string;
+    result: AnalysisResult | null;
+}
+
+const SldOverlay: React.FC<SldOverlayProps> = ({
+    vlOverlay, actionViewMode,
+    onOverlayClose, onOverlaySldTabChange,
+    n1Diagram, actionDiagram,
+    selectedBranch, result,
+}) => {
+    const overlayBodyRef = useRef<HTMLDivElement>(null);
+    const [overlayPos, setOverlayPos] = useState({ x: 16, y: 16 });
+    const [overlayTransform, setOverlayTransform] = useState({ scale: 1, tx: 0, ty: 0 });
+
+    // Apply / clear delta flow colors on the SLD whenever svg/metadata, tab, or mode changes.
+    //
+    // pypowsybl SLD SVG structure (GraphMetadata):
+    //   feederNodes[].id         → SVG element ID of the feeder <g> (sld-top-feeder / sld-bottom-feeder)
+    //   feederNodes[].equipmentId → network equipment ID (matches flow_deltas keys)
+    //
+    // The cell ancestor (sld-extern-cell) wraps BOTH the feeder symbol AND the connecting
+    // wire (sld-wire), so we must apply the delta class one level up to color the full branch.
+    useEffect(() => {
+        const container = overlayBodyRef.current;
+        if (!container) return;
+
+        // Clear any previously applied SLD delta classes
+        const SLD_DELTA_CLASSES = [
+            'sld-delta-positive', 'sld-delta-negative', 'sld-delta-grey',
+            'sld-delta-text-positive', 'sld-delta-text-negative', 'sld-delta-text-grey'
+        ];
+        container.querySelectorAll(SLD_DELTA_CLASSES.map(c => '.' + c).join(','))
+            .forEach(el => el.classList.remove(...SLD_DELTA_CLASSES));
+
+        // Restore original text labels
+        container.querySelectorAll('[data-original-text]').forEach(el => {
+            el.textContent = el.getAttribute('data-original-text');
+            el.removeAttribute('data-original-text');
+        });
+
+        // Restore flipped arrow directions (sld-in ↔ sld-out)
+        container.querySelectorAll('[data-arrow-flipped]').forEach(el => {
+            if (el.classList.contains('sld-in')) {
+                el.classList.replace('sld-in', 'sld-out');
+            } else if (el.classList.contains('sld-out')) {
+                el.classList.replace('sld-out', 'sld-in');
+            }
+            el.removeAttribute('data-arrow-flipped');
+        });
+
+        if (!vlOverlay.svg || actionViewMode !== 'delta') return;
+
+        // Choose deltas based on the SLD tab being shown
+        const flowDeltas = vlOverlay.flow_deltas;
+        const reactiveDeltas = vlOverlay.reactive_flow_deltas;
+        const assetDeltas = vlOverlay.asset_deltas;
+
+        if (!flowDeltas && !assetDeltas) return;
+
+        // Build equipmentId → [svgId, ...] multimap from SLD metadata.
+        // pypowsybl SLD metadata uses 'nodes' (for lines, transformers, breakers,
+        // bus-bar sections) and 'feederInfos' (for ARROW_ACTIVE/ARROW_REACTIVE).
+        // Older versions may use 'feederNodes' instead.
+        const equipIdToSvgIds = new Map<string, string[]>();
+        if (vlOverlay.sldMetadata) {
+            try {
+                const meta = JSON.parse(vlOverlay.sldMetadata) as {
+                    nodes?: SldFeederNode[];
+                    feederInfos?: SldFeederNode[];
+                    feederNodes?: SldFeederNode[];
+                };
+                // Collect entries from all possible metadata arrays
+                const sources = [
+                    ...(meta.nodes ?? []),
+                    ...(meta.feederInfos ?? []),
+                    ...(meta.feederNodes ?? []),
+                ];
+                for (const fn of sources) {
+                    if (fn.equipmentId && fn.id) {
+                        const ids = equipIdToSvgIds.get(fn.equipmentId) ?? [];
+                        ids.push(fn.id);
+                        equipIdToSvgIds.set(fn.equipmentId, ids);
+                    }
+                }
+            } catch {
+                // metadata parse failed — fall through to substring fallback
+            }
+        }
+
+        // Quick lookup of all elements by SVG id
+        const elMap = new Map<string, Element>();
+        container.querySelectorAll('[id]').forEach(el => elMap.set(el.id, el));
+
+        /**
+         * Look up an SVG element by ID, trying the exact ID first and then
+         * common sanitization variants (pypowsybl sometimes replaces dots with
+         * underscores in SVG element IDs while preserving the original in metadata).
+         */
+        const lookupById = (svgId: string): Element | undefined =>
+            elMap.get(svgId)
+            ?? elMap.get(svgId.replace(/\./g, '_'))   // dots → underscores
+            ?? elMap.get(svgId.replace(/_/g, '.'));    // underscores → dots
+
+        /**
+         * Look up a key in a Record, trying the exact key first and then
+         * dot↔underscore variants.  pypowsybl may sanitize dots in equipment
+         * IDs differently between get_lines() (used for flow_deltas keys) and
+         * SLD metadata (used for equipmentId).
+         */
+        const lookupDelta = <T,>(rec: Record<string, T> | null | undefined, key: string): T | undefined => {
+            if (!rec) return undefined;
+            const exact = rec[key];
+            if (exact !== undefined) return exact;
+            const dotted = key.replace(/_/g, '.');
+            if (dotted !== key && rec[dotted] !== undefined) return rec[dotted];
+            const underscored = key.replace(/\./g, '_');
+            if (underscored !== key && rec[underscored] !== undefined) return rec[underscored];
+            return undefined;
+        };
+
+        const applyTextDelta = (label: Element, val: string) => {
+            if (!label.hasAttribute('data-original-text')) {
+                label.setAttribute('data-original-text', label.textContent || '');
+            }
+            label.textContent = `\u0394 ${val}`;
+        };
+
+        const fmtDelta = (v: number) => v >= 0 ? `+${v.toFixed(1)}` : v.toFixed(1);
+
+        /** Walk up from a feeder element to the enclosing cell ancestor. */
+        const walkUpToCell = (feederEl: Element): Element => {
+            let cellEl: Element = feederEl;
+            let cur: Element | null = feederEl.parentElement;
+            while (cur && cur !== container) {
+                if (cur.classList.contains('sld-extern-cell') ||
+                    cur.classList.contains('sld-intern-cell') ||
+                    cur.classList.contains('sld-shunt-cell')) {
+                    cellEl = cur;
+                    break;
+                }
+                cur = cur.parentElement;
+            }
+            return cellEl;
+        };
+
+        /**
+         * Find the cell ancestor element for a given equipment ID.
+         * Tries the metadata-based SVG IDs first (with sanitization variants),
+         * then falls back to substring matching against all element IDs.
+         */
+        const findCellEl = (equipId: string): Element | null => {
+            let feederEl: Element | undefined;
+            // Try exact key, then dot↔underscore variants in metadata map
+            const svgIds = equipIdToSvgIds.get(equipId)
+                ?? equipIdToSvgIds.get(equipId.replace(/\./g, '_'))
+                ?? equipIdToSvgIds.get(equipId.replace(/_/g, '.'));
+            if (svgIds) {
+                for (const svgId of svgIds) {
+                    feederEl = lookupById(svgId);
+                    if (feederEl) break;
+                }
+            }
+            if (!feederEl) {
+                // Substring fallback: also try with dots replaced by underscores
+                const sanitized = equipId.replace(/\./g, '_');
+                for (const [eid, el] of elMap) {
+                    if (eid.includes(equipId) || (sanitized !== equipId && eid.includes(sanitized))) {
+                        feederEl = el;
+                        break;
+                    }
+                }
+            }
+            if (!feederEl) return null;
+            return walkUpToCell(feederEl);
+        };
+
+        /** Replace the first numeric label in a query result with a delta string. */
+        const replaceFirstNumericLabel = (labels: NodeListOf<Element>, val: string) => {
+            for (const label of Array.from(labels)) {
+                if (/^-?\d+(\.\d+)?$/.test((label.textContent || '').trim())) {
+                    applyTextDelta(label, val);
+                    return;
+                }
+            }
+        };
+
+        // Helper to apply P & Q delta text to a cell element
+        const applyPQLabels = (cellEl: Element, pStr: string, qStr: string | null) => {
+            let pLabels = cellEl.querySelectorAll('.sld-active-power .sld-label');
+            if (pLabels.length === 0) pLabels = cellEl.querySelectorAll('.sld-label');
+            replaceFirstNumericLabel(pLabels, pStr);
+
+            if (qStr !== null) {
+                const qLabels = cellEl.querySelectorAll('.sld-reactive-power .sld-label');
+                if (qLabels.length > 0) {
+                    replaceFirstNumericLabel(qLabels, qStr);
+                } else {
+                    // Fallback: P label already replaced (no longer numeric),
+                    // so the first remaining numeric label IS the Q label.
+                    replaceFirstNumericLabel(cellEl.querySelectorAll('.sld-label'), qStr);
+                }
+            }
+        };
+
+        /** Flip arrow direction within a specific power-type scope.
+         *  pypowsybl SLD SVGs use .sld-in / .sld-out classes to control
+         *  which arrow (.sld-arrow-in / .sld-arrow-out) is visible.
+         *  P and Q arrows are flipped independently by scoping to
+         *  .sld-active-power or .sld-reactive-power. */
+        const flipArrows = (cellEl: Element, scopeClass: string) => {
+            const sel = `.${scopeClass} .sld-in, .${scopeClass} .sld-out, .${scopeClass}.sld-in, .${scopeClass}.sld-out`;
+            cellEl.querySelectorAll(sel).forEach(el => {
+                if (el.hasAttribute('data-arrow-flipped')) return;
+                if (el.classList.contains('sld-in')) {
+                    el.classList.replace('sld-in', 'sld-out');
+                } else {
+                    el.classList.replace('sld-out', 'sld-in');
+                }
+                el.setAttribute('data-arrow-flipped', '1');
+            });
+        };
+
+        // Iterate ALL feeders from SLD metadata so we process branches, loads,
+        // and generators — not just equipment IDs found in flow_deltas.
+        const processedEquipIds = new Set<string>();
+
+        for (const [equipId, svgIds] of equipIdToSvgIds) {
+            // Find the first matching feeder element across all SVG IDs for this equipment
+            let feederEl: Element | undefined;
+            for (const svgId of svgIds) {
+                feederEl = lookupById(svgId);
+                if (feederEl) break;
+            }
+            if (!feederEl) continue;
+
+            const cellEl = walkUpToCell(feederEl);
+
+            // Check branch (line/transformer) deltas first
+            const branchDelta = lookupDelta(flowDeltas, equipId);
+            if (branchDelta) {
+                cellEl.classList.add(`sld-delta-${branchDelta.category}`);
+
+                const pStr = fmtDelta(branchDelta.delta);
+                const qDelta = lookupDelta(reactiveDeltas, equipId);
+                const qStr = qDelta !== undefined ? fmtDelta(qDelta.delta) : null;
+
+                // Always apply labels to ensure original flows are replaced
+                applyPQLabels(cellEl, pStr, qStr);
+
+                // Apply specific category classes to the labels instead of the cell
+                let pLabels = cellEl.querySelectorAll('.sld-active-power .sld-label');
+                if (pLabels.length === 0) pLabels = cellEl.querySelectorAll('.sld-label');
+                pLabels.forEach(l => l.classList.add(`sld-delta-text-${branchDelta.category}`));
+
+                if (qDelta) {
+                    cellEl.querySelectorAll('.sld-reactive-power .sld-label').forEach(l => l.classList.add(`sld-delta-text-${qDelta.category}`));
+                }
+
+                // Flip P and Q arrows independently (regardless of category)
+                if (branchDelta.flip_arrow) {
+                    flipArrows(cellEl, 'sld-active-power');
+                }
+                if (qDelta && qDelta.flip_arrow) {
+                    flipArrows(cellEl, 'sld-reactive-power');
+                }
+                processedEquipIds.add(equipId);
+                continue;
+            }
+
+            // Check asset (load/generator) deltas
+            const assetDelta = lookupDelta(assetDeltas, equipId);
+            if (assetDelta) {
+                cellEl.classList.add(`sld-delta-${assetDelta.category}`);
+                const pStr = fmtDelta(assetDelta.delta_p);
+                const qStr = fmtDelta(assetDelta.delta_q);
+                // Always apply labels to ensure original flows are replaced
+                applyPQLabels(cellEl, pStr, qStr);
+
+                // Use independent categories for labels if available
+                const catP = assetDelta.category_p || assetDelta.category;
+                const catQ = assetDelta.category_q || assetDelta.category;
+
+                cellEl.querySelectorAll('.sld-active-power .sld-label').forEach(l => l.classList.add(`sld-delta-text-${catP}`));
+                cellEl.querySelectorAll('.sld-reactive-power .sld-label').forEach(l => l.classList.add(`sld-delta-text-${catQ}`));
+
+                processedEquipIds.add(equipId);
+            }
+        }
+
+        // Helper: check if an equipment ID (or a dot↔underscore variant) was
+        // already processed in the metadata-based loop above.
+        const isProcessed = (id: string): boolean =>
+            processedEquipIds.has(id)
+            || processedEquipIds.has(id.replace(/\./g, '_'))
+            || processedEquipIds.has(id.replace(/_/g, '.'));
+
+        // Fallback: process any flow_deltas / asset_deltas keys not found
+        // via metadata (in case metadata was incomplete or parse failed).
+        for (const [equipId, delta] of Object.entries(flowDeltas ?? {})) {
+            if (isProcessed(equipId)) continue;
+            const cellEl = findCellEl(equipId);
+            if (!cellEl) continue;
+            const pStr = fmtDelta(delta.delta);
+            const qDelta = lookupDelta(reactiveDeltas, equipId);
+            const qStr = qDelta !== undefined ? fmtDelta(qDelta.delta) : null;
+
+            // Always apply labels to ensure original flows are replaced
+            applyPQLabels(cellEl, pStr, qStr);
+
+            let pLabels = cellEl.querySelectorAll('.sld-active-power .sld-label');
+            if (pLabels.length === 0) pLabels = cellEl.querySelectorAll('.sld-label');
+            pLabels.forEach(l => l.classList.add(`sld-delta-text-${delta.category}`));
+
+            if (qDelta) {
+                cellEl.querySelectorAll('.sld-reactive-power .sld-label').forEach(l => l.classList.add(`sld-delta-text-${qDelta.category}`));
+            }
+
+            if (delta.flip_arrow) flipArrows(cellEl, 'sld-active-power');
+            const qDeltaForFlip = lookupDelta(reactiveDeltas, equipId);
+            if (qDeltaForFlip && qDeltaForFlip.flip_arrow) flipArrows(cellEl, 'sld-reactive-power');
+        }
+        for (const [equipId, assetDelta] of Object.entries(assetDeltas ?? {})) {
+            if (isProcessed(equipId)) continue;
+            if (lookupDelta(flowDeltas, equipId) !== undefined) continue;
+            const cellEl = findCellEl(equipId);
+            if (!cellEl) continue;
+            if (assetDelta.category !== 'grey' || (assetDelta.category_q && assetDelta.category_q !== 'grey')) {
+                const pStr = fmtDelta(assetDelta.delta_p);
+                const qStr = fmtDelta(assetDelta.delta_q);
+                // Always apply labels to ensure original flows are replaced
+                applyPQLabels(cellEl, pStr, qStr);
+
+                const catP = assetDelta.category_p || assetDelta.category;
+                const catQ = assetDelta.category_q || assetDelta.category;
+
+                cellEl.querySelectorAll('.sld-active-power .sld-label').forEach(l => l.classList.add(`sld-delta-text-${catP}`));
+                cellEl.querySelectorAll('.sld-reactive-power .sld-label').forEach(l => l.classList.add(`sld-delta-text-${catQ}`));
+            }
+        }
+    }, [vlOverlay.svg, vlOverlay.sldMetadata, vlOverlay.tab, actionViewMode,
+    vlOverlay.flow_deltas, vlOverlay.reactive_flow_deltas, vlOverlay.asset_deltas]);
+
+    // ===== Highlight impacted assets on the SLD =====
+    // Uses a clone-behind technique: clones target elements and inserts them as
+    // direct siblings (before the original) so they naturally track the original
+    // during pan/zoom without needing CTM-based repositioning.
+    useEffect(() => {
+        const container = overlayBodyRef.current;
+        if (!container || !vlOverlay.svg) return;
+
+        // Remove previous highlight clones
+        container.querySelectorAll('.sld-highlight-clone').forEach(el => el.remove());
+
+        // Remove previous "-original" classes from assets
+        const ORIGINAL_CLASSES = [
+            'sld-highlight-contingency-original',
+            'sld-highlight-action-original',
+            'sld-highlight-breaker-original',
+            'sld-highlight-overloaded-original'
+        ];
+        container.querySelectorAll(ORIGINAL_CLASSES.map(c => '.' + c).join(','))
+            .forEach(el => el.classList.remove(...ORIGINAL_CLASSES));
+
+        // Build equipmentId → SVG element ID map from SLD metadata
+        const equipIdToSvgIds = new Map<string, string[]>();
+        if (vlOverlay.sldMetadata) {
+            try {
+                const meta = JSON.parse(vlOverlay.sldMetadata) as {
+                    nodes?: SldFeederNode[];
+                    feederInfos?: SldFeederNode[];
+                    feederNodes?: SldFeederNode[];
+                };
+                const sources = [
+                    ...(meta.nodes ?? []),
+                    ...(meta.feederInfos ?? []),
+                    ...(meta.feederNodes ?? []),
+                ];
+                for (const fn of sources) {
+                    if (fn.equipmentId && fn.id) {
+                        const ids = equipIdToSvgIds.get(fn.equipmentId) ?? [];
+                        ids.push(fn.id);
+                        equipIdToSvgIds.set(fn.equipmentId, ids);
+                    }
+                }
+            } catch {
+                // metadata parse failed
+            }
+        }
+        if (equipIdToSvgIds.size === 0) return;
+
+        const svg = container.querySelector('svg');
+        if (!svg) return;
+
+        // Build element ID → DOM element map
+        const elMap = new Map<string, Element>();
+        container.querySelectorAll('[id]').forEach(el => elMap.set(el.id, el));
+
+        const lookupById = (svgId: string): Element | undefined =>
+            elMap.get(svgId)
+            ?? elMap.get(svgId.replace(/\./g, '_'))
+            ?? elMap.get(svgId.replace(/_/g, '.'));
+
+        const walkUpToCell = (el: Element): Element => {
+            let cellEl: Element = el;
+            let cur: Element | null = el.parentElement;
+            while (cur && cur !== container) {
+                if (cur.classList.contains('sld-extern-cell') ||
+                    cur.classList.contains('sld-intern-cell') ||
+                    cur.classList.contains('sld-shunt-cell')) {
+                    cellEl = cur;
+                    break;
+                }
+                cur = cur.parentElement;
+            }
+            return cellEl;
+        };
+
+        const findCellForEquipment = (equipId: string): Element | null => {
+            const svgIds = equipIdToSvgIds.get(equipId)
+                ?? equipIdToSvgIds.get(equipId.replace(/\./g, '_'))
+                ?? equipIdToSvgIds.get(equipId.replace(/_/g, '.'));
+            if (svgIds) {
+                for (const svgId of svgIds) {
+                    const el = lookupById(svgId);
+                    if (el) return walkUpToCell(el);
+                }
+            }
+            for (const [metaEquipId, metaSvgIds] of equipIdToSvgIds) {
+                if (metaEquipId.includes(equipId) || equipId.includes(metaEquipId)) {
+                    for (const svgId of metaSvgIds) {
+                        const el = lookupById(svgId);
+                        if (el) return walkUpToCell(el);
+                    }
+                }
+            }
+            const sanitized = equipId.replace(/\./g, '_');
+            for (const [eid, el] of elMap) {
+                if (eid.includes(equipId) || (sanitized !== equipId && eid.includes(sanitized))) {
+                    return walkUpToCell(el);
+                }
+            }
+            return null;
+        };
+
+        // Clone an element and insert it right before the original in the DOM.
+        // SVG renders in document order, so the clone (earlier) appears behind
+        // the original (later). Being a sibling in the same parent group means
+        // the clone moves with the original during pan/zoom — no CTM needed.
+        const cloneHighlight = (el: Element, highlightClass: string) => {
+            const clone = el.cloneNode(true) as SVGGraphicsElement;
+            clone.removeAttribute('id');
+            clone.querySelectorAll('[id]').forEach(child => child.removeAttribute('id'));
+            clone.classList.add('sld-highlight-clone', highlightClass);
+            el.parentNode?.insertBefore(clone, el);
+            el.classList.add(highlightClass + '-original');
+        };
+
+        const highlightedCells = new Set<Element>();
+
+        const actionId = vlOverlay.actionId;
+        const actionDetail: ActionDetail | undefined =
+            (actionId && result?.actions) ? result.actions[actionId] : undefined;
+
+        // --- Contingency highlight (N-1 and action tabs) ---
+        if (vlOverlay.tab !== 'n' && selectedBranch) {
+            const cell = findCellForEquipment(selectedBranch);
+            if (cell) {
+                cloneHighlight(cell, 'sld-highlight-contingency');
+                highlightedCells.add(cell);
+            }
+        }
+
+        // --- Action target highlights (action tab) ---
+        if (vlOverlay.tab === 'action' && actionDetail) {
+            const topo = actionDetail.action_topology;
+            const isCoupling = isCouplingAction(actionId, actionDetail.description_unitaire);
+
+            if (topo) {
+                // For coupling actions: only highlight the breaker/switch, not branches.
+                // For other actions: highlight affected lines/gens/loads.
+                if (!isCoupling) {
+                    const targetEquipIds = new Set<string>();
+                    for (const id of Object.keys(topo.lines_ex_bus || {})) targetEquipIds.add(id);
+                    for (const id of Object.keys(topo.lines_or_bus || {})) targetEquipIds.add(id);
+                    for (const id of Object.keys(topo.gens_bus || {})) targetEquipIds.add(id);
+                    for (const id of Object.keys(topo.loads_bus || {})) targetEquipIds.add(id);
+                    for (const id of Object.keys(topo.pst_tap || {})) targetEquipIds.add(id);
+                    for (const id of Object.keys(topo.loads_p || {})) targetEquipIds.add(id);
+                    for (const id of Object.keys(topo.gens_p || {})) targetEquipIds.add(id);
+
+                    for (const equipId of targetEquipIds) {
+                        const cell = findCellForEquipment(equipId);
+                        if (cell && !highlightedCells.has(cell)) {
+                            cloneHighlight(cell, 'sld-highlight-action');
+                            highlightedCells.add(cell);
+                        }
+                    }
+                }
+
+                // Highlight affected breakers/switches.
+                // Prefer changed_switches from SLD response (robust N-1 vs action comparison)
+                // over action_topology.switches (may be empty for grid2op actions).
+                const changedSwitches = vlOverlay.changed_switches;
+                const topoSwitches = topo.switches as Record<string, unknown> | undefined;
+                const switchSource = changedSwitches && Object.keys(changedSwitches).length > 0
+                    ? changedSwitches
+                    : topoSwitches;
+                // Coupling action breakers highlighted yellow (same as action targets);
+                // regular breaker/switch actions use purple.
+                const breakerClass = isCoupling ? 'sld-highlight-action' : 'sld-highlight-breaker';
+                if (switchSource) {
+                    for (const switchId of Object.keys(switchSource)) {
+                        const cell = findCellForEquipment(switchId);
+                        if (cell && !highlightedCells.has(cell)) {
+                            cloneHighlight(cell, breakerClass);
+                            highlightedCells.add(cell);
+                        }
+                    }
+                }
+            }
+
+            // Fallback: parse description for breaker/switch names
+            const breakerClass = isCouplingAction(actionId, actionDetail.description_unitaire)
+                ? 'sld-highlight-action' : 'sld-highlight-breaker';
+            const desc = actionDetail.description_unitaire;
+            if (desc && highlightedCells.size === 0) {
+                const ocMatch = desc.match(/OC\s+'([^']+)'/);
+                if (ocMatch) {
+                    const ocName = ocMatch[1].replace(/\s+DJ_OC$/, '');
+                    const cell = findCellForEquipment(ocName);
+                    if (cell && !highlightedCells.has(cell)) {
+                        cloneHighlight(cell, breakerClass);
+                        highlightedCells.add(cell);
+                    }
+                }
+                const lineMatch = desc.match(/(?:Ouverture|Fermeture)\s+(\S+)\s+DJ_OC/);
+                if (lineMatch) {
+                    const cell = findCellForEquipment(lineMatch[1]);
+                    if (cell && !highlightedCells.has(cell)) {
+                        cloneHighlight(cell, breakerClass);
+                        highlightedCells.add(cell);
+                    }
+                }
+            }
+        }
+
+        // --- Overloaded lines (N-1 tab only) ---
+        if (vlOverlay.tab === 'n-1') {
+            const overloadedLines = result?.lines_overloaded;
+            if (overloadedLines && overloadedLines.length > 0) {
+                for (const lineId of overloadedLines) {
+                    const cell = findCellForEquipment(lineId);
+                    if (cell && !highlightedCells.has(cell)) {
+                        cloneHighlight(cell, 'sld-highlight-overloaded');
+                        highlightedCells.add(cell);
+                    }
+                }
+            }
+        }
+
+    }, [vlOverlay.svg, vlOverlay.sldMetadata, vlOverlay.tab, vlOverlay.actionId,
+    vlOverlay.changed_switches, selectedBranch, result]);
+
+    // Non-passive wheel zoom on overlay body
+    useEffect(() => {
+        const el = overlayBodyRef.current;
+        if (!el) return;
+        const onWheel = (e: WheelEvent) => {
+            e.preventDefault();
+            const factor = e.deltaY < 0 ? 1.15 : 1 / 1.15;
+            const rect = el.getBoundingClientRect();
+            const cx = e.clientX - rect.left;
+            const cy = e.clientY - rect.top;
+            setOverlayTransform(prev => {
+                const s = Math.max(0.1, Math.min(10, prev.scale * factor));
+                return { scale: s, tx: cx - (cx - prev.tx) * (s / prev.scale), ty: cy - (cy - prev.ty) * (s / prev.scale) };
+            });
+        };
+        el.addEventListener('wheel', onWheel, { passive: false });
+        return () => el.removeEventListener('wheel', onWheel);
+    }, []);
+
+    const startOverlayDrag = (e: React.MouseEvent) => {
+        if (e.button !== 0) return;
+        e.preventDefault();
+        const x0 = e.clientX, y0 = e.clientY;
+        const px0 = overlayPos.x, py0 = overlayPos.y;
+        const onMove = (ev: MouseEvent) => setOverlayPos({ x: px0 + ev.clientX - x0, y: py0 + ev.clientY - y0 });
+        const onUp = () => { window.removeEventListener('mousemove', onMove); window.removeEventListener('mouseup', onUp); };
+        window.addEventListener('mousemove', onMove);
+        window.addEventListener('mouseup', onUp);
+    };
+
+    const startOverlayPan = (e: React.MouseEvent) => {
+        if (e.button !== 0) return;
+        e.preventDefault();
+        const x0 = e.clientX, y0 = e.clientY;
+        const tx0 = overlayTransform.tx, ty0 = overlayTransform.ty;
+        const onMove = (ev: MouseEvent) => setOverlayTransform(prev => ({ ...prev, tx: tx0 + ev.clientX - x0, ty: ty0 + ev.clientY - y0 }));
+        const onUp = () => { window.removeEventListener('mousemove', onMove); window.removeEventListener('mouseup', onUp); };
+        window.addEventListener('mousemove', onMove);
+        window.addEventListener('mouseup', onUp);
+    };
+
+    return (
+        <div style={{
+            position: 'absolute', top: overlayPos.y + 'px', left: overlayPos.x + 'px',
+            width: '440px', height: '420px', minWidth: '220px', minHeight: '150px',
+            background: 'white', border: '1px solid #ccc', borderRadius: '8px',
+            boxShadow: '0 4px 24px rgba(0,0,0,0.22)', zIndex: 45,
+            display: 'flex', flexDirection: 'column', overflow: 'hidden',
+            resize: 'both', boxSizing: 'border-box',
+        }}>
+            {/* Header — drag handle */}
+            <div
+                onMouseDown={startOverlayDrag}
+                style={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', padding: '6px 10px', background: '#f0faf4', borderBottom: '1px solid #d1fae5', flexShrink: 0, cursor: 'move', userSelect: 'none' }}
+            >
+                <div style={{ display: 'flex', flexDirection: 'column', gap: '4px' }}>
+                    <div style={{ display: 'flex', alignItems: 'center', gap: '6px' }}>
+                        <span style={{ fontSize: '13px', fontWeight: 600, color: '#065f46' }}>{vlOverlay.vlName}</span>
+                        {/* Mode indicator — shows which Flow vs Impact mode was active when overlay opened */}
+                        <span style={{
+                            fontSize: '10px', fontWeight: 700, padding: '1px 6px', borderRadius: '10px',
+                            background: actionViewMode === 'delta' ? '#dbeafe' : '#f3f4f6',
+                            color: actionViewMode === 'delta' ? '#1d4ed8' : '#374151',
+                        }}>
+                            {actionViewMode === 'delta' ? 'Impacts' : 'Flows'}
+                        </span>
+                    </div>
+                    <div style={{ display: 'flex', gap: '4px' }}>
+                        {(['n', 'n-1', 'action'] as SldTab[]).filter(tabMode => {
+                            if (tabMode === 'n-1') return !!n1Diagram;
+                            if (tabMode === 'action') return !!actionDiagram;
+                            return true; // always show N
+                        }).map(tabMode => (
+                            <button
+                                key={tabMode}
+                                onMouseDown={e => e.stopPropagation()}
+                                onClick={(e) => { e.stopPropagation(); onOverlaySldTabChange(tabMode); }}
+                                style={{
+                                    background: vlOverlay.tab === tabMode ? '#059669' : '#e5e7eb',
+                                    color: vlOverlay.tab === tabMode ? 'white' : '#374151',
+                                    border: 'none', borderRadius: '4px', padding: '2px 8px',
+                                    fontSize: '11px', fontWeight: vlOverlay.tab === tabMode ? 'bold' : 'normal',
+                                    cursor: vlOverlay.loading ? 'wait' : 'pointer',
+                                }}
+                            >
+                                {tabMode.toUpperCase()}
+                            </button>
+                        ))}
+                    </div>
+                </div>
+                <button
+                    onMouseDown={e => e.stopPropagation()}
+                    onClick={(e) => { e.stopPropagation(); onOverlayClose(); }}
+                    style={{ background: 'none', border: 'none', cursor: 'pointer', fontSize: '16px', color: '#666', lineHeight: 1, padding: '0 2px' }}
+                    title="Close"
+                >✕</button>
+            </div>
+            {/* Body — pan/zoom canvas */}
+            <div
+                ref={overlayBodyRef}
+                style={{ flex: 1, overflow: 'hidden', minHeight: 0, cursor: 'grab', userSelect: 'none' }}
+                onMouseDown={startOverlayPan}
+            >
+                {vlOverlay.loading && (
+                    <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'center', height: '100%', color: '#999', fontSize: '13px' }}>
+                        Generating diagram…
+                    </div>
+                )}
+                {vlOverlay.error && (
+                    <div style={{ padding: '12px', color: '#dc3545', fontSize: '12px' }}>{vlOverlay.error}</div>
+                )}
+                {vlOverlay.svg && (
+                    <div style={{
+                        transformOrigin: '0 0',
+                        transform: `translate(${overlayTransform.tx}px,${overlayTransform.ty}px) scale(${overlayTransform.scale})`,
+                        padding: '4px',
+                    }} dangerouslySetInnerHTML={{ __html: vlOverlay.svg }} />
+                )}
+            </div>
+        </div>
+    );
+};
+
+export default SldOverlay;
